Tests/LibWeb/Text/expected/HTML/HTMLElement-offsetFoo-for-split-inline-element.txt

b.offsetTop: 8
b.offsetLeft: 8
b.offsetWidth: 784
b.offsetHeight: 51
